requirements.txt
pandas
jupyter

# Google Cloud
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib

## app
dash
dash_bootstrap_components
dash_uploader
psycopg2
pypdf

## llm
groq

## others
python-dotenv